refactor: rename metrics_pipeline → parquet_pipeline, update CODEOWNERS
The actors under quickwit-indexing/src/actors/metrics_pipeline/ are
the parquet-based ingest+compaction pipeline. They currently process
metrics, but logs and traces will move onto the same pipeline in
later phases — the metrics-specific name has become misleading.

Pure rename:
- git mv metrics_pipeline → parquet_pipeline
- update mod declaration + import paths (metrics_pipeline:: →
  parquet_pipeline::)
- update CODEOWNERS path

Function names that incidentally contain "metrics_pipeline"
(spawn_metrics_pipeline, spawn_log_or_metrics_pipeline,
test_metrics_pipeline_e2e) are NOT renamed in this commit — they
describe pipeline kind ("the metrics-flavored pipeline") rather than
the module, and would benefit from their own pass that also addresses
the eventual log/trace generalization.

CODEOWNERS also adds /docs/internals/ to the byoc-metrics path list:
the architecture and verification docs there are domain-aligned with
quickwit-parquet-engine + quickwit-dst (already in the list) and
should be approvable by the same team.
