+++ quickwit/quickwit-indexing/src/actors/parquet_pipeline/indexing_service_impl.rs
